+++ CHANGELOG.md
@@ -9,6 +9,7 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 ### Changed
 
 ### Fixed
+- Unread counter displays a huge number of unread posts
 
 # Releases
 ## [26.1.0] - 2025-08-02